Refactor tests to reuse `intellijPlatformDependency` logic and ensure consistent local IntelliJ Platform setup.
